feat(scripts): add _nk.py NkeyProvider ABC + Slice-1 stubs for #1017 (T11)
